Add_food_entry update
Что было сделано:
- переделан язык разметки, для корректного отображения на маленьких экранах;
- обновлено отображение КБЖУ, расчитанного по введенному кол-ву грамм;
- исправлен баг с невозможностью введения вещественного кол-ва грамм;
- задано ограничение кол-ва цифр после запятой у БЖУ и грамм в FoodEntry.

+++ total_control_django/calculator_app/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 
 class FoodEntryForm(forms.Form):
-    grams = forms.IntegerField(
+    grams = forms.FloatField(
         label='Граммы',
         min_value=1,
         max_value=15000,
